+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   Transactions form was reset when creating a new category or subcategory from the form.
 -   Fix bug that avoids deleting a category.
 -   Fix backup listing to not consider macOS metadata files (.\_\*.json) as valid backups.
 -   Next occurrence calculation for scheduled transactions list was not considering modifications properly, leading to incorrect next occurrence dates.